Replace the probing strategy for selecting the read buffer
Previously StripedBuffer used a per-thread hash code that was randomly
reset when contention was detected. In the optimized case this used a
private field inside of Thread, threadLocalRandomProbe. This is derived
from Java's Striped64 which uses this technique.

As the JVM becomes more restrictive in what can be accessed, the code
was allowed to fallback to a thread local value. This degraded the
read throughput by ~14M ops/s on my laptop. While satisfactory as
over 100M ops/s, the strategy needed to be reevaluated.

@DougLea suggested using an incremental hash based on the thread's id.
While this loses the perfect hash rebalancing, the benchmarks show an
equivalent throughput as the randomized probe. This may benefit thanks
to Caffeine being allowed to drop events when it exceeds a retry
threshold, whereas LongAdder must retry until the counter is
incremented. We retain the lazy allocation growth and other tricks from
Striped64 and drop the unsafe field access.

+++ caffeine/src/main/java/com/github/benmanes/caffeine/cache/BoundedBuffer.java
@@ -103,13 +103,13 @@ final class BoundedBuffer<E> extends StripedBuffer<E> {
     }
 
     @Override
-    public int reads() {
-      return (int) readCounter;
+    public long reads() {
+      return readCounter;
     }
 
     @Override
-    public int writes() {
-      return (int) writeCounter;
+    public long writes() {
+      return writeCounter;
     }
   }
 }
