+++ caffeine/src/test/java/com/github/benmanes/caffeine/cache/buffer/FastFlowBuffer.java
@@ -85,13 +85,13 @@ final class FastFlowBuffer<E> extends FastFlowHeader.ReadAndWriteCounterRef<E> {
   }
 
   @Override
-  public int reads() {
-    return (int) readCounter;
+  public long reads() {
+    return readCounter;
   }
 
   @Override
-  public int writes() {
-    return (int) writeCounter;
+  public long writes() {
+    return writeCounter;
   }
 }
